Using some C++0x features - Core.h: using C++11 static_assert for staticAssert when possible - fixed warnings in C++11 mode of GCC:   - concatenation of strings - can't use "str1""str2", but use "str1" "str2" (needs a space between quotes) - see changes with enabled whitespace differences   - RGBA macro - use 'unsigned' instead of 'int' to avoid warning about conversion of negative int to unsigned - added "-std=c++0x" option to GCC (required for static_assert)

+++ Core/GlWindow.h
@@ -76,10 +76,10 @@ extern bool GShowDebugInfo;
 #undef RGB					// defined in windows
 
 // constant colors
-#define RGBA(r,g,b,a)		((int)((r)*255) | ((int)((g)*255)<<8) | ((int)((b)*255)<<16) | ((int)((a)*255)<<24))
+#define RGBA(r,g,b,a)		( (unsigned)((r)*255) | ((unsigned)((g)*255)<<8) | ((unsigned)((b)*255)<<16) | ((unsigned)((a)*255)<<24) )
 #define RGB(r,g,b)			RGBA(r,g,b,1)
-#define RGB255(r,g,b)		((r) | ((g)<<8) | ((b)<<16) | (255<<24))
-#define RGBA255(r,g,b,a)	((r) | ((g)<<8) | ((b)<<16) | ((a)<<24))
+#define RGB255(r,g,b)		( (unsigned) ((r) | ((g)<<8) | ((b)<<16) | (255<<24)) )
+#define RGBA255(r,g,b,a)	( (r) | ((g)<<8) | ((b)<<16) | ((a)<<24) )
 
 // computed colors
 //?? make as methods; or - constructor of CColor
